% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_seg.R
\name{common.seg}
\alias{common.seg}
\title{Segment factor-driven common component}
\usage{
common.seg(
  x,
  demean = TRUE,
  G.seq = NULL,
  thr = NULL,
  tt.by = floor(2 * log(dim(x)[2])),
  eta = 0.5
)
}
\arguments{
\item{x}{input time series matrix, with each row representing a variable}

\item{demean}{whether to de-mean the input \code{x} row-wise}

\item{G.seq}{a sequence of integers [TO DO]}

\item{thr}{}

\item{tt.by}{}

\item{eta}{}
}
\value{
a list containing the following fields:
\item{est.cp}{ }
\item{G.seq}{ }
\item{thr}{ }
\item{est.cp.list}{ a list containing the following fields:
\itemize{
\item{\code{cp}}{ }
\item{\code{G}}{ }
\item{\code{ll}}{ }
\item{\code{norm.stat}}{ }
\item{\code{stat}}{ }
}}
\item{mean.x}{ if \code{center = TRUE}, returns a vector containing row-wise sample means of \code{x}; if \code{center = FALSE}, returns a vector of zeros}
}
\description{
Segment factor-driven common component
}
\details{
See Algorithm 1 of Cho, Eckley, Fearnhead and Maeng (2022) for further details.
}
\references{
H. Cho, I. Eckley, P. Fearnhead and H. Maeng (2022) High-dimensional time series segmentation via factor-adjusted vector autoregressive modelling. arXiv preprint arXiv: TODO
}
