% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sim.data}
\alias{sim.data}
\title{Simulate a piecewise stationary, factor-adjusted VAR process}
\usage{
sim.data(
  n,
  p,
  q = 2,
  cp.common = round(n * 1:3/4),
  den.common = 0.5,
  type.common = c("ma", "ar")[1],
  ma.order = 2,
  cp.idio = round(n * c(3, 5)/8),
  d = 1,
  size.idio = 1,
  do.scale = TRUE,
  seed = NULL
)
}
\description{
Simulate a piecewise stationary, factor-adjusted VAR process
}
