% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio_seg.R
\name{yw.cv}
\alias{yw.cv}
\title{Cross validation for factor-adjusted VAR estimation}
\usage{
yw.cv(
  xx,
  lambda.max = NULL,
  var.order = 1,
  n.folds = 1,
  path.length = 10,
  q = 0,
  kern.const = 4
)
}
\description{
Cross validation for factor-adjusted VAR estimation
}
\keyword{internal}
