% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_seg.R
\name{common.seg}
\alias{common.seg}
\title{Segment factor-driven common component}
\usage{
common.seg(
  x,
  center = TRUE,
  G.seq = NULL,
  thr = NULL,
  alpha = 0.1,
  tt.by = floor(2 * log(dim(x)[2])),
  eta = 0.5
)
}
\arguments{
\item{x}{input time series matrix, with each row representing a variable}

\item{center}{whether to de-mean the input \code{x} row-wise}

\item{G.seq}{an integer vector of bandwidth; if \code{G.seq = NULL}, a default choice \code{G.seq = round(n * 1/c(10, 8, 6, 4))} is used}

\item{thr}{a vector of thresholds which is of the same length as \code{G.seq}; if \code{thr = NULL}, a default choice based on numerical experiments is used}

\item{alpha}{used when \code{thr = NULL} for default threshold selection, which relates to the quantile used in numerical experiments; \code{alpha = 0.05} and \code{alpha = 0.1} are supported}

\item{tt.by}{an integer specifying the grid over which the test statistic is computed, which is \code{round(seq(G, dim(x)[2] - G, by = tt.by))} for each bandwidth \code{G}}

\item{eta}{a constant between \code{0} and \code{1}; each local maximiser of the test statistic within its \code{eta * G}-environment for the common component is deemed as a change point estimator. Also the bottom-up merging across the multiple bandwidths \code{G.seq} depends on this parameter}
}
\value{
a list containing the following fields:
\item{est.cp}{ a matrix containing the change point estimators in the first column and the finest bandwidth at which each is detected in the second column }
\item{G.seq}{ an integer vector of bandwidths }
\item{thr}{ a vector of thresholds which is of the same length as \code{G.seq} }
\item{est.cp.list}{ a list containing the following fields:
\itemize{
\item{\code{cp}}{ change point estimators }
\item{\code{G}}{ bandwidth }
\item{\code{ll}}{ kernel window size for spectral density estimation }
\item{\code{norm.stat}}{ a matrix containing test statistic values at Fourier frequencies }
\item{\code{stat}}{ a vector containing test statistic values across multiple frequencies }
}}
\item{mean.x}{ if \code{center = TRUE}, returns a vector containing row-wise sample means of \code{x}; if \code{center = FALSE}, returns a vector of zeros}
}
\description{
Segment factor-driven common component
}
\details{
See Algorithm 1 of Cho, Eckley, Fearnhead and Maeng (2022) for further details.
}
\examples{
\dontrun{
out <- sim.data(n = 2000, p = 50, q = 2, d = 1,
cp.common = 1:3/4, den.common = .5, type.common = 'ma', 
cp.idio = c(3, 5)/8, seed = 123)
cs <- common.seg(out$x)
cs$est.cp
}
}
\references{
Cho, H., Eckley, I., Fearnhead, P. & Maeng, H. (2022) High-dimensional time series segmentation via factor-adjusted vector autoregressive modelling. arXiv preprint arXiv:2204.02724
}
