% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idio_seg.R
\name{idio.seg}
\alias{idio.seg}
\title{Segment idiosyncratic VAR process}
\usage{
idio.seg(
  x,
  demean = TRUE,
  common.seg.out,
  q = NULL,
  d = 1,
  G.seq = NULL,
  thr = NULL,
  eta = 0.5,
  cv.args = list(path.length = 10, n.folds = 1, do.cv = FALSE)
)
}
\arguments{
\item{x}{input time series matrix, with each row representing a variable}

\item{demean}{whether to de-mean the input \code{x} row-wise}

\item{q}{an integer specifying the number of factors. If \code{q = NULL}, the factor number is estimated by an information criterion-based approach of Hallin and Liška (2007) for each segment}

\item{d}{an integer specifying the VAR order}

\item{G.seq}{a sequence of integers [TO DO]}

\item{thr}{}

\item{eta}{}

\item{cv.args}{a list specifying the tuning parameters required for Dantzig selector tuning parameter selection via cross-validation:
\itemize{
   \item{\code{n.folds}}{ number of folds}
   \item{\code{path.length}}{ number of regularisation parameter values to consider; a sequence is generated automatically based in this value}
   \item{\code{do.cv}}{ if \code{do.cv = FALSE}, a fixed value is selected from a sequence of 10 values chosen in a data-driven way}
}}
}
\value{
a list containing the following fields:
\item{est.cp}{ }
\item{G.seq}{ }
\item{thr}{ }
\item{est.cp.list}{ a list containing the following fields:
\itemize{
\item{\code{cp}}{ }
\item{\code{G}}{ }
\item{\code{ll}}{ }
\item{\code{norm.stat}}{ }
\item{\code{stat}}{ }
}}
\item{mean.x}{ if \code{center = TRUE}, returns a vector containing row-wise sample means of \code{x}; if \code{center = FALSE}, returns a vector of zeros}
}
\description{
Segment idiosyncratic VAR process
}
\details{
See Algorithm 2 of Cho, Eckley, Fearnhead and Maeng (2022) for further details.
}
\references{
H. Cho, I. Eckley, P. Fearnhead and H. Maeng (2022) High-dimensional time series segmentation via factor-adjusted vector autoregressive modelling. arXiv preprint arXiv: TODO

Hallin, M. & Liška, R. (2007) Determining the number of factors in the general dynamic factor model. Journal of the American Statistical Association, 102(478), 603--617.
}
