% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sim.data}
\alias{sim.data}
\title{Simulate a piecewise stationary factor-adjusted VAR process}
\usage{
sim.data(
  n,
  p,
  q = 2,
  d = 1,
  cp.common = c(1:3)/4,
  den.common = 0.5,
  type.common = c("ma", "ar"),
  ma.order = 2,
  cp.idio = c(3, 5)/8,
  size.idio = 1,
  do.scale = TRUE,
  seed = NULL
)
}
\arguments{
\item{n}{sample size}

\item{p}{number of variables}

\item{q}{number of dynamic factors}

\item{d}{VAR order}

\item{cp.common}{a vector specifying the re-scaled locations of the change points between \code{0} and \code{1} in the common component; possible to set \code{cp.common = c()} (no change point)}

\item{den.common}{a value between \code{0} and \code{1} specifying the cross-sectional density of each change point}

\item{type.common}{if \code{type.common = 'ma'}, factors are loaded as innovations of a moving average process with order \code{ma.order}; if \code{type.common = 'ar'}, factors are loaded as innovations of an autoregressive process of order \code{1}}

\item{ma.order}{order of the factor-driven moving average process; used when \code{type.common = 'ma'}}

\item{cp.idio}{a vector specifying the re-scaled locations of the change points between \code{0} and \code{1} in the idiosyncratic component; possible to set \code{cp.idio = c()} (no change point)}

\item{size.idio}{at each change point, each of VAR parameter matrices has its sign changed and is multiplied by \code{size.idio}}

\item{do.scale}{if \code{do.scale = TRUE}, each variable of the common component is scaled to have the same sample variance as the corresponding idiosyncratic variable}

\item{seed}{an integer setting the seed of the random number generator}
}
\value{
a list containing
\item{x}{ generated piecewise stationary factor-adjusted vector autoregressive process }
\item{xi}{ generated piecewise stationary vector autoregressive process }
\item{A.list}{ a list containing the VAR parameter matrices over the segments }
\item{cp.common}{ input parameter }
\item{cp.idio}{ input parameter }
}
\description{
Generate time series used in the simulation studies of Cho, Eckley, Fearnhead and Maeng (2022) for further details.
}
\examples{
out <- sim.data(n = 2000, p = 100, q = 2, d = 1,
cp.common = 1:3/4, den.common = .5, type.common = 'ma', 
cp.idio = c(3, 5)/8, seed = 123)
}
\references{
H. Cho, I. Eckley, P. Fearnhead and H. Maeng (2022) High-dimensional time series segmentation via factor-adjusted vector autoregressive modelling. arXiv preprint arXiv: TODO
}
